consolidate case test into their test files
